Performance
When having some byzantine code, for example, having an literal array with 10000 elements, each one of these elements is
adding the array node to process it. Therefore, we end up processing the same node, in the same way, 10000 times.

On some byzantine code, it gave an improvement of 170s -> 85s (twice!). Sane assumption would be, that the bigger the
codebase, the bigger chance that there's a file/script like that, and we do want to support big codebases.